backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
websockets==12.0
google-genai>=1.0.0
httpx==0.27.0
youtube-transcript-api==0.6.2
pydantic==2.6.1
aiofiles==23.2.1
